Back-End/whyEat/requirements.txt
asgiref==3.2.7
astroid==2.3.3
attrs==19.3.0
autopep8==1.5.1
backcall==0.1.0
bleach==3.1.4
branca==0.3.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.5
django-cors-headers==3.2.1
django-rest-knox==4.1.0
django-rest-swagger==2.1.2
djangorestframework==3.10.3
dnspython==1.16.0
drf-yasg==1.17.1
entrypoints==0.3
folium==0.10.1
idna==2.9
importlib-metadata==1.6.0
inflection==0.4.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itypes==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
mysql-connector-python==8.0.19
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.5
notebook==6.0.3
numpy==1.17.3
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.3
pandas==0.25.3
pandocfilters==1.4.2
parso==0.6.2
pickleshare==0.7.5
Pillow==7.1.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.6.1
pycodestyle==2.5.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
PyMySQL==0.9.3
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
simplejson==3.17.0
six==1.14.0
social-auth-core==3.3.3
SQLAlchemy==1.3.16
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.1.0
